backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
greenlet==3.0.3

# LTI 1.3
PyLTI1p3==2.0.0
jwcrypto==1.5.6
cryptography==42.0.7

# HTTP / sessions
httpx==0.27.0
itsdangerous==2.2.0
python-multipart==0.0.9
starlette==0.37.2

# AI providers
google-genai>=1.0.0
ollama==0.2.1

# RAG
pypdf>=4.0.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
